requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.9.0

# DNS and networking
dnspython>=2.4.0

# Configuration management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: Report generation
jinja2>=3.1.2